Exercise-Snowflake-26-Clustering/requirements.txt
snowflake-connector-python[pandas]==3.4.0
numpy==1.24.3
pandas==2.0.3
pyarrow==10.0.1
faker==19.6.2
python-dotenv==1.0.0
